chime: prebuilts: workaround a prebuilt app limitation
[@eldainosor];
Looks like currently, prebuilt apps are being modified by dex and jni to optimize their use. The bad thing is: This break it's structure, causing Package Manager to not take our prebuilt apps as valid due to the following error: No APK Signature Scheme v2 signature in package. (or if you install it via adb, it errors out with is signed using APK Signature Scheme v2, but no such signature was found. Signature stripped?)
This is caused due to the apk structure being changed and not finding the right memory block for the signature. (maybe, not sure about this).
This workaround replaces the repacked apk with the original, avoiding the optimization but solving the issue.

Signed-off-by: itsHanibee <itshanibee@gmail.com>

Former-commit-id: 9f174a6422625003948684cf75d76a288250d1c1